fix(ci): cover shared tests and retire stale agent workflows
