+++ .github/workflows/ci.yml
@@ -130,6 +130,47 @@ jobs:
           } >> "$GITHUB_STEP_SUMMARY"
 
   # ---------------------------------------------------------------------------
+  # Job 1b: Shared contracts package
+  # ---------------------------------------------------------------------------
+  test-shared:
+    name: Test Shared Contracts
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - name: Setup Node 22
+        uses: actions/setup-node@53b83947a5a98c8d113130e565377fae1a50d02f # v6.3.0
+        with:
+          node-version-file: '.nvmrc'
+          package-manager-cache: false
+
+      - name: Enable corepack and install pnpm
+        run: corepack enable && corepack prepare pnpm@10.32.1 --activate
+
+      - name: Get pnpm store path
+        id: pnpm-cache
+        run: echo "store=$(pnpm store path)" >> "$GITHUB_OUTPUT"
+
+      - name: Cache pnpm store
+        uses: actions/cache@668228422ae6a00e4ad889ee87cd7109ec5666a7 # v5
+        with:
+          path: ${{ steps.pnpm-cache.outputs.store }}
+          key: pnpm-store-${{ runner.os }}-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            pnpm-store-${{ runner.os }}-
+
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+
+      - name: Type-check shared contracts
+        run: pnpm --filter @sentinel/shared lint
+
+      - name: Run shared contracts tests
+        run: pnpm --filter @sentinel/shared test
+
+  # ---------------------------------------------------------------------------
   # Job 2: Python quant engine
   # ---------------------------------------------------------------------------
   test-engine:
@@ -228,7 +269,7 @@ jobs:
       (github.event_name == 'push' && github.ref == 'refs/heads/main') ||
       (github.event_name == 'pull_request' &&
         (github.base_ref == 'main' || startsWith(github.base_ref, 'release/')))
-    needs: [verify-commit-signatures, test-web, test-engine, test-agents]
+    needs: [verify-commit-signatures, test-web, test-shared, test-engine, test-agents]
     runs-on: ubuntu-latest
     timeout-minutes: 10
