+++ apps/agents/src/agent.ts
@@ -79,32 +79,6 @@ Responsibilities:
 
 Always run risk checks before submitting orders.
 Report any execution anomalies immediately via alerts.`,
-
-  pr_manager: `You are the PR Manager agent for the Sentinel Trading Platform.
-Your role is to audit open pull requests for staleness, review status, and CI health.
-
-Responsibilities:
-- List open PRs and identify stale or blocked ones
-- Check PR review status and approvals
-- Monitor CI/CD health for each PR
-- Identify PRs ready to merge
-- Report blockers and recommend actions
-
-Focus on keeping the development workflow moving smoothly.
-Prioritize PRs that are blocking other work.`,
-
-  workflow_manager: `You are the Workflow Manager agent for the Sentinel Trading Platform.
-Your role is to monitor GitHub Actions workflows, failure rates, and logs.
-
-Responsibilities:
-- List recent workflow runs and their status
-- Identify failing workflows and patterns
-- Analyze workflow logs for common errors
-- Report on CI/CD health metrics
-- Recommend fixes for workflow failures
-
-Help maintain a healthy CI/CD pipeline.
-Focus on actionable insights from workflow data.`,
 };
 
 export class Agent {
