+++ apps/agents/src/error-handling.ts
@@ -159,45 +159,3 @@ export class CircuitBreaker {
     return this.state;
   }
 }
-
-/**
- * Deduplication cache for identical concurrent requests.
- * Prevents duplicate API calls within the same cycle.
- */
-export class RequestDeduplicator {
-  private cache = new Map<string, { promise: Promise<unknown>; expiresAt: number }>();
-
-  /**
-   * Get cached result or execute function.
-   */
-  async dedupe<T>(key: string, fn: () => Promise<T>, ttlMs: number = 60_000): Promise<T> {
-    const cached = this.cache.get(key);
-
-    if (cached && cached.expiresAt > Date.now()) {
-      return cached.promise as Promise<T>;
-    }
-
-    const promise = fn();
-
-    this.cache.set(key, {
-      promise,
-      expiresAt: Date.now() + ttlMs,
-    });
-
-    try {
-      const result = await promise;
-      return result;
-    } catch (error) {
-      this.cache.delete(key);
-      throw error;
-    }
-  }
-
-  clear(): void {
-    this.cache.clear();
-  }
-
-  size(): number {
-    return this.cache.size;
-  }
-}
