+++ apps/web/src/components/advisor/memory-timeline.tsx
@@ -2,6 +2,7 @@
 
 import { Badge } from '@/components/ui/badge';
 import { cn } from '@/lib/utils';
+import { formatRelativeTime } from '@/lib/format-relative-time';
 import type { AdvisorMemoryEvent, MemoryEventType } from '@sentinel/shared';
 import { MEMORY_EVENT_LABELS } from '@sentinel/shared';
 import { Brain, Check, Pencil, X, Trash2, RotateCcw, Clock, User } from 'lucide-react';
@@ -34,18 +35,6 @@ const eventColors: Record<MemoryEventType, string> = {
   preference_restored: 'text-primary',
   preference_auto_expired: 'text-muted-foreground',
 };
-
-function formatRelativeTime(dateStr: string): string {
-  const diff = Date.now() - new Date(dateStr).getTime();
-  const mins = Math.floor(diff / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 7) return `${days}d ago`;
-  return new Date(dateStr).toLocaleDateString();
-}
 
 export function MemoryTimeline({ events, isLoading, className }: MemoryTimelineProps) {
   if (isLoading) {
