+++ docs/ai/state/project-state.md
@@ -55,6 +55,7 @@ Before editing files:
 
 | Ticket                   | Status | Owner  | Branch/Worktree                               | Scope                                                                                                                                                                                                                                                               | Validation                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       | Last Updated | Notes                                                                                                                                                                                                                                                                                                 |
 | ------------------------ | ------ | ------ | --------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------ | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| OPS-2026-04-08A          | done   | Codex  | `fix/ci-agents-ui-contract-audit`             | `.github/workflows/ci.yml`, `apps/agents/src/**`, `apps/agents/workflows/**`, `apps/web/src/components/advisor/**`, `apps/web/src/lib/format-relative-time.ts`, `docs/workflows-and-agents.md`, `docs/ai/state/project-state.md`                                    | `pnpm lint`; `pnpm --filter agents test`; `pnpm --filter @sentinel/shared test`; `pnpm --filter web test --run tests/pages/advisor.test.tsx`; `pnpm --filter web build`                                                                                                                                                                                                                                                                                                                                                          | 2026-04-08   | Added missing shared-contract CI test job, retired PR/Workflow manager agent roles and workflows, consolidated advisor relative-time UI pattern, fixed nested-button semantics, and aligned architecture docs/contracts.                                                                              |
 | T-OPS02                  | done   | Codex  | `chore/railway-supabase-audit`                | `scripts/validate-railway-supabase-env.mjs`, `scripts/validate-railway-supabase-env.test.mjs`, `docs/runbooks/production.md`, `docs/ai/commands.md`, `docs/ai/state/project-state.md`                                                                               | `node --test scripts/validate-railway-supabase-env.test.mjs`; `node scripts/validate-railway-supabase-env.mjs --profile=web --production --project-ref=luwyjfwauljwsfsnwiqb --env-file=/tmp/web.env`; `node scripts/validate-railway-supabase-env.mjs --profile=engine --project-ref=luwyjfwauljwsfsnwiqb --env-file=/tmp/engine.env`; `node scripts/validate-railway-supabase-env.mjs --profile=agents --production --project-ref=luwyjfwauljwsfsnwiqb --require-private-engine --env-file=/tmp/agents.env`; `git diff --check` | 2026-04-07   | Added deeper validator hardening with env-file support, key/secret quality checks, import-safe module design, and node:test regression coverage to prevent contract drift.                                                                                                                            |
 | T-OPS01                  | done   | Codex  | `chore/railway-supabase-audit`                | `docs/runbooks/production.md`, `scripts/validate-railway-supabase-env.mjs`, `docs/ai/commands.md`, `docs/ai/state/project-state.md`                                                                                                                                 | `node scripts/validate-railway-supabase-env.mjs --profile=web --production --project-ref=luwyjfwauljwsfsnwiqb`; `node scripts/validate-railway-supabase-env.mjs --profile=engine --project-ref=luwyjfwauljwsfsnwiqb`; `node scripts/validate-railway-supabase-env.mjs --profile=agents --production --project-ref=luwyjfwauljwsfsnwiqb --require-private-engine`; `git diff --check`                                                                                                                                             | 2026-04-07   | Added executable Railway↔Supabase env validation checks and wired production runbook + commands docs to enforce project-ref matching, private networking, and production URL hygiene.                                                                                                                 |
 | T0                       | done   | Codex  | `work`                                        | `docs/ai/roadmaps/2026-03-20-deployment-readme-roadmap.md`, `docs/ai/agent-ops.md`, `docs/ai/state/project-state.md`                                                                                                                                                | `git diff --check`                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               | 2026-03-20   | Seeded roadmap, operating guide, and state ledger requested by user.                                                                                                                                                                                                                                  |
